fix(mobile): make tests runnable in bun without React Native runtime
Extract mergeRecord into pure lib/merge-record.ts so tests don't pull
in react-native-sse or React hooks. Remove expo-constants dependency
from config.ts (use env var for backend URL). Add mobile tests to root
test script.
