+++ packages/mobile/lib/config.ts
@@ -1,8 +1,6 @@
-import Constants from 'expo-constants'
-
-const devBackendUrl = Constants.expoConfig?.extra?.backendUrl ?? 'http://localhost:3000'
-
-export const API_BASE_URL: string = __DEV__ ? devBackendUrl : 'https://api.bartools.wtf'
+// Backend URL — override via BARTOOLS_API_URL env var for testing
+export const API_BASE_URL: string =
+  process.env.BARTOOLS_API_URL ?? 'http://localhost:3000'
 
 // Hardcoded until auth is implemented
 export const DEFAULT_USER_ID = '00000000-0000-0000-0000-000000000001'
